+++ src/Main/Features/DbWindow.h
@@ -8,8 +8,8 @@
 
 #include <Actions.h>
 #include <AgentData.h>
-#include <Callbacks.h>
 #include <GuiUtils.h>
+#include <HelperCallbacks.h>
 #include <HelperUw.h>
 #include <Move.h>
 #include <PlayerData.h>
