+++ src/Main/Features/EmoWindow.cpp
@@ -23,7 +23,7 @@
 #include <Actions.h>
 #include <GuiUtils.h>
 #include <Helper.h>
-#include <HelperPlayer.h>
+#include <HelperAgents.h>
 #include <HelperUw.h>
 #include <HelperUwPos.h>
 #include <Logger.h>
