+++ src/Types/Move.cpp
@@ -5,7 +5,7 @@
 
 #include <AgentData.h>
 #include <Helper.h>
-#include <HelperPlayer.h>
+#include <HelperAgents.h>
 #include <HelperUw.h>
 #include <HelperUwPos.h>
 #include <PlayerData.h>
